fix: global page controller 추가, 수정
